edited userservice and recommendation service

+++ user-service/src/main/java/com/stackroute/kafka/domain/User.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class User {
 
     @Id
-    private long uid;
+    private int uid;
 
     private String name;
